requirements.txt
uvicorn = "^0.14.0"
peewee = "^3.14.4"
fastapi = "^0.66.0"
requests = "^2.26.0"
Pillow = "^8.3.1"